+++ classes/output/restore_details.php
@@ -16,6 +16,7 @@
 
 namespace tool_vault\output;
 
+use tool_vault\api;
 use tool_vault\constants;
 use tool_vault\local\checks\check_base;
 use tool_vault\local\helpers\ui;
@@ -77,6 +78,7 @@ class restore_details implements \templatable {
             'restoreid' => $this->restore->id,
             'resumeurl' => $this->restore->can_resume() ?
                 \tool_vault\local\uiactions\restore_resume::url(['id' => $this->restore->id])->out(false) : null,
+            'vaultfaqurl' => api::get_frontend_url().'/faq',
         ];
 
         if ($this->restore->status == constants::STATUS_INPROGRESS) {